AI-teams-controller-public/experiments/TTS/google-cloud/pyproject.toml
[project]
name = "google-cloud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-texttospeech>=2.33.0",
]
